#1499-Fix 'Duplicate Code' issue in multiple files
